test: move computed-attribute uniqueness tests to node create-flow file
The replication tests drive create_node() directly — they exercise the
production create path, not the NodeGroupedUniquenessConstraint class in
isolation. Move them to backend/tests/component/core/node/ alongside
test_template_pool_allocation.py, which is the existing home for
create_node integration tests. Drop the now-redundant constraint-level
test that overlapped with the create-flow coverage.
